% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programme.R
\name{add_case_detection_cost}
\alias{add_case_detection_cost}
\title{Proactive case detection cost}
\usage{
add_case_detection_cost(x)
}
\arguments{
\item{x}{output}
}
\value{
output with case detection costs
}
\description{
Proactive case detection cost
}
