% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{add_facility_costs}
\alias{add_facility_costs}
\title{Facility costs}
\usage{
add_facility_costs(x, iso3c)
}
\arguments{
\item{x}{output}

\item{iso3c}{Country iso3c code}
}
\value{
output with facility costs
}
\description{
Facility costs
}
