% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{add_facility_visits}
\alias{add_facility_visits}
\title{Number of facility visits.}
\usage{
add_facility_visits(x)
}
\arguments{
\item{x}{output}
}
\value{
output with facility visits
}
\description{
Assumed some non-severe are in the private sector and all severe visits in public sector
}
