% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_irs}
\alias{add_irs}
\title{IRS commodities}
\usage{
add_irs(x)
}
\arguments{
\item{x}{output}
}
\value{
output with number of people protected and number of households sprayed
}
\description{
IRS commodities
}
