% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add missing columns}
\usage{
add_missing_columns(x, species = "pf")
}
\arguments{
\item{x}{malariasimulation output}

\item{species}{species code. Can be falicparum "pf" or vivax "pf".}
}
\value{
data.frame with all required columns.
}
\description{
Ensure that the minimum set of required columns is present, even if a simulation
run isn't performed (i.e. when there is no transmission of a species in a
location, or no implementation of an intervention in a location).
}
