% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_nets}
\alias{add_nets}
\title{Add nets}
\usage{
add_nets(x, iso3c, country_ur = FALSE)
}
\arguments{
\item{x}{output}

\item{iso3c}{Country ISO3c code}

\item{country_ur}{Use country-level usage rate estimates}
}
\value{
output with net numbers
}
\description{
Add nets
}
