% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{add_pf_columns}
\alias{add_pf_columns}
\title{Add missing columns to pf model output}
\usage{
add_pf_columns(
  x,
  col = c("n_rtss_epi_dose_1", "n_rtss_epi_dose_2", "n_rtss_epi_dose_3",
    "n_rtss_epi_booster_1", "n_pmc_treated", "n_smc_treated")
)
}
\arguments{
\item{x}{output}

\item{col}{Vector of required column names}
}
\value{
output with all required columns
}
\description{
New columns are assigned 0
}
