% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_pmc}
\alias{add_pmc}
\title{Add PMC}
\usage{
add_pmc(x, modelled_population_size, doses_per_child = 3)
}
\arguments{
\item{x}{output}

\item{modelled_population_size}{modelled population size}

\item{doses_per_child}{Number of PMC doses per child per year}
}
\value{
output with pmc doses and children protected
}
\description{
Add PMC
}
