% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{add_rates}
\alias{add_rates}
\title{Converts modelled quantities of age-disaggregated cases and severe cases
to rates}
\usage{
add_rates(x, modelled_population_size)
}
\arguments{
\item{x}{Model output}

\item{modelled_population_size}{Modelled population size}
}
\value{
Output with rates
}
\description{
Converts modelled quantities of age-disaggregated cases and severe cases
to rates
}
