% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_rtss}
\alias{add_rtss}
\title{Add RTS,S}
\usage{
add_rtss(x, modelled_population_size, rtss_doses_per_child = 4)
}
\arguments{
\item{x}{output}

\item{modelled_population_size}{modelled population size}

\item{rtss_doses_per_child}{Number of doses per child}
}
\value{
output with rtss doses and rtss protected
}
\description{
Add RTS,S
}
