% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{add_rtss_total}
\alias{add_rtss_total}
\title{Add total number of RTS,S doses}
\usage{
add_rtss_total(x)
}
\arguments{
\item{x}{output}
}
\value{
output with n_rtss_doses
}
\description{
Add total number of RTS,S doses
}
