% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{add_smc}
\alias{add_smc}
\title{Add SMC}
\usage{
add_smc(x, modelled_population_size)
}
\arguments{
\item{x}{output}

\item{modelled_population_size}{modelled population size}
}
\value{
output with pmc doses and children protected
}
\description{
Add SMC
}
