% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programme.R
\name{add_surveillance_cost}
\alias{add_surveillance_cost}
\title{Surveillance cost}
\usage{
add_surveillance_cost(x)
}
\arguments{
\item{x}{output}
}
\value{
output with surveillance costs
}
\description{
Surveillance cost
}
