% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{add_tx}
\alias{add_tx}
\title{Treatment and diagnostic commodities}
\usage{
add_tx(x)
}
\arguments{
\item{x}{output}
}
\value{
output with treatment and diagnostic commodities
}
\description{
Treatment and diagnostic commodities
}
