% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{add_tx_internal}
\alias{add_tx_internal}
\title{Add treatment commodities}
\usage{
add_tx_internal(x, suffix = NULL)
}
\arguments{
\item{x}{model output}

\item{suffix}{add "_lower" or "_upper" for uncertainty}
}
\value{
output with treatment commodities
}
\description{
Add treatment commodities
}
