% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{add_year}
\alias{add_year}
\title{Add year column based on model timestep and baseline year}
\usage{
add_year(x, baseline_year = 2000)
}
\arguments{
\item{x}{model output}

\item{baseline_year}{start year}
}
\value{
output with year column
}
\description{
Add year column based on model timestep and baseline year
}
