% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{age_mid_point}
\alias{age_mid_point}
\title{Add average age in range (assuming exponential demography)}
\usage{
age_mid_point(x)
}
\arguments{
\item{x}{output}
}
\value{
output with age average
}
\description{
Add average age in range (assuming exponential demography)
}
