% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{beta_coverage_ci}
\alias{beta_coverage_ci}
\title{Beta distributed uncertainty for coverage-related commodity outputs}
\usage{
beta_coverage_ci(coverage, q, average_sample)
}
\arguments{
\item{coverage}{Coverage}

\item{q}{Quantile}

\item{average_sample}{Average sample size for coverage estimate. These can
be informed via the denominators from DHS surveys (for example, median sample
size from most recent surveys for IRS = 9720, for children with fever taking
an act = 1473).}
}
\value{
Uncertainty estimates on coverage
}
\description{
Beta distributed uncertainty for coverage-related commodity outputs
}
