% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{empty_pf}
\alias{empty_pf}
\title{Empty pf output}
\usage{
empty_pf(timesteps, baseline_year = 2000)
}
\arguments{
\item{timesteps}{Model timesteps}

\item{baseline_year}{Baseline year}
}
\value{
Empty output
}
\description{
To be used when pf eir = 0
}
