% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{empty_pv}
\alias{empty_pv}
\title{Empty pv output}
\usage{
empty_pv(timesteps, baseline_year = 2000)
}
\arguments{
\item{timesteps}{Model timesteps}

\item{baseline_year}{Baseline year}
}
\value{
Empty output
}
\description{
To be used when pv eir = 0
}
