% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{epi_uncertainty_approximation}
\alias{epi_uncertainty_approximation}
\title{Uncertainty approximation for epi outcomes}
\usage{
epi_uncertainty_approximation(n, q, cv)
}
\arguments{
\item{n}{N number of cases/deaths}

\item{q}{Quantile}

\item{cv}{Uncertainty SD scaler, cases = 0.227, deaths = 0.265}
}
\description{
Uncertainty approximation for epi outcomes
}
