% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{exp_mid}
\alias{exp_mid}
\title{Approximate mean age within an age bracket assuming exponential
age distribution}
\usage{
exp_mid(l, u, average_age = 19 * 365)
}
\arguments{
\item{l}{age bracket min}

\item{u}{age bracket max}

\item{average_age}{Average age in days}
}
\value{
output with average age column
}
\description{
Approximate mean age within an age bracket assuming exponential
age distribution
}
