% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{format_age}
\alias{format_age}
\title{Wrangle raw age-group output}
\usage{
format_age(x, population_size)
}
\arguments{
\item{x}{model output}

\item{population_size}{Modelled population size}
}
\description{
Wrangle raw age-group output
}
