% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{format_epi}
\alias{format_epi}
\title{Wrangle raw epi output}
\usage{
format_epi(x, species)
}
\arguments{
\item{x}{Model output}

\item{species}{Species "pf" or "pv"}
}
\description{
Wrangle raw epi output
}
