% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{format_interventions}
\alias{format_interventions}
\title{Wrangle raw intervention output}
\usage{
format_interventions(x)
}
\arguments{
\item{x}{Model output}
}
\description{
Wrangle raw intervention output
}
