% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{format_prev}
\alias{format_prev}
\title{Wrangle raw prevalence output}
\usage{
format_prev(x, species)
}
\arguments{
\item{x}{Model output}

\item{species}{Species "pf" or "pv"}
}
\description{
Wrangle raw prevalence output
}
