% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_age_aggregated}
\alias{get_age_aggregated}
\title{Age aggregated output}
\usage{
get_age_aggregated(
  age_disaggregated,
  iso3c,
  interventions,
  scenario,
  run,
  model_population
)
}
\arguments{
\item{age_disaggregated}{age disaggregated output}

\item{iso3c}{COutnry iso3c code}

\item{interventions}{site interventions}

\item{scenario}{Scenario name}

\item{run}{Run name}

\item{model_population}{Modelled population size}
}
\value{
Age aggregated output
}
\description{
Age aggregated output
}
