% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_age_disaggregated}
\alias{get_age_disaggregated}
\title{Age disaggregated}
\usage{
get_age_disaggregated(pf, pv, iso3c, population)
}
\arguments{
\item{pf}{pf output}

\item{pv}{pv output}

\item{iso3c}{iso3c code}

\item{population}{site population data}
}
\value{
Age disaggregated output
}
\description{
Age disaggregated
}
