% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_pf}
\alias{get_pf}
\title{Pf output}
\usage{
get_pf(model_output, interventions, population, model_population)
}
\arguments{
\item{model_output}{model output}

\item{interventions}{site interventions}

\item{population}{site population}

\item{model_population}{Modelled population size}
}
\value{
pf output
}
\description{
Pf output
}
