% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_pv}
\alias{get_pv}
\title{Pv output}
\usage{
get_pv(model_output, interventions, population, model_population)
}
\arguments{
\item{model_output}{model output}

\item{interventions}{site interventions}

\item{population}{site population}

\item{model_population}{Modelled population size}
}
\value{
pv output
}
\description{
Pv output
}
