% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_interventions}
\alias{link_interventions}
\title{Link intervention inputs}
\usage{
link_interventions(x, interventions)
}
\arguments{
\item{x}{output}

\item{interventions}{interventions}
}
\value{
output linked to model intervention inputs
}
\description{
Link intervention inputs
}
