% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_pop}
\alias{link_pop}
\title{Link PAR and adjust for age proportions}
\usage{
link_pop(x, pop)
}
\arguments{
\item{x}{output}

\item{pop}{pop data frame (cols: year and par)}
}
\value{
output with par
}
\description{
Link PAR and adjust for age proportions
}
