% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linking.R
\name{link_tx}
\alias{link_tx}
\title{Link treatment coverage and add coverage uncertainty}
\usage{
link_tx(x, tx)
}
\arguments{
\item{x}{output}

\item{tx}{treatment data.frame (cols: year, tx_cov)}
}
\value{
output with treatment coverage
}
\description{
Link treatment coverage and add coverage uncertainty
}
