% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{rates}
\alias{rates}
\title{Scale outputs for modelled population size}
\usage{
rates(x, modelled_population)
}
\arguments{
\item{x}{Model outputs}

\item{modelled_population}{modelled population size}
}
\value{
output with prop (age), inc and severe inc
}
\description{
Scale outputs for modelled population size
}
