% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_raw.R
\name{rename_cols}
\alias{rename_cols}
\title{Add appropriate species suffix to column names}
\usage{
rename_cols(x, suffix)
}
\arguments{
\item{x}{output}

\item{suffix}{species suffix (pf or pv)}
}
\value{
renamed output
}
\description{
Add appropriate species suffix to column names
}
