% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf.R
\name{wrangle_pf}
\alias{wrangle_pf}
\title{Wrangle pf model output}
\usage{
wrangle_pf(x)
}
\arguments{
\item{x}{model output}
}
\value{
Long, age-disaggregated pf data
}
\description{
Collapses output to annual and converts to long wrt age disaggregation
}
