% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{wrangle_pv}
\alias{wrangle_pv}
\title{Wrangle pv model output}
\usage{
wrangle_pv(x)
}
\arguments{
\item{x}{model output}
}
\value{
Long, age-disaggregated pv data
}
\description{
Collapses output to annual and converts to long wrt age disaggregation
}
