% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{yld}
\alias{yld}
\title{Years of life with disability}
\usage{
yld(
  x,
  episode_length = 0.01375,
  severe_episode_length = 0.04795,
  weight1 = 0.211,
  weight2 = 0.195,
  weight3 = 0.172,
  severe_weight = 0.6
)
}
\arguments{
\item{x}{output}

\item{episode_length}{Average length of clinical episode}

\item{severe_episode_length}{Average length of severe episode}

\item{weight1}{Disability weight age group 1}

\item{weight2}{Disability weight age group 2}

\item{weight3}{Disability weight age group 3}

\item{severe_weight}{Disability weight severe malaria}
}
\value{
output with yld
}
\description{
Weights from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4772264/}{Gunda _et al_, 2016}.
Currently using pf values.
}
