% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{yll}
\alias{yll}
\title{Years of life lost}
\usage{
yll(x, life_expectancy)
}
\arguments{
\item{x}{output}

\item{life_expectancy}{Life expectancy}
}
\value{
output with yll
}
\description{
An approximation of years of life lost (within discretised age groups)
}
