% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_tx.R
\name{add_tx}
\alias{add_tx}
\title{Add treatment commodities.}
\usage{
add_tx(x)
}
\arguments{
\item{x}{output}
}
\value{
output with treatment commodities
}
\description{
Adds treatment commodities and uncertainty based of uncertainty in
treatment coverage, cases and severe cases.
}
