% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{mortality_rate}
\alias{mortality_rate}
\title{Add mortality rate}
\usage{
mortality_rate(x, scaler, treatment_scaler = 0.5)
}
\arguments{
\item{x}{output}

\item{scaler}{Scaler for severe cases to deaths. Have used
original fitted scaler  of 0.215 for falciparum. For vivax can use a scaler
of 0.003 (Ahmed Rahimi, et al. (2014)).}

\item{treatment_scaler}{Impact of treatment coverage on scaling.}
}
\value{
output with mortality rates
}
\description{
Add mortality rate
}
